Symbol.cry: remove net nops
It now becomes clear that we cons up Small::T instances only to unwrap
them again, and never do anything with Object::T instances at all. To
be fair the latter's partly because object-sized symbols aren't really
supported yet. But a lot of what was in place (and obscured by all the
submodules and imports of separate files and functor applications)
wasn't actually doing anything.