Fold all the value type cryptol logic into one file.
This includes minimal changes to the logic, and minimal changes to
everything else to make things go again.

Having one file instead of one per type is less obviously the way to
go than it was before cleaning out all the functors, but I still think
it's what we want, given that in the long run there's quite a few more
value types we don't yet support.