Rework the value logic, part 1: type tags.
Note for the record: the whole value rework (along with the
corresponding changes to soroban.saw and other files) was done at once
(over several days) going direct from the original cryptol code with
submodules and functors and all. All the incremental cleanup in the
past commits today (going back to the first functor application
changes) is retrospective; also this and the next few commits are just
doing one part of Values.cry at a time for clarity.

I'm committing it in pieces for two reasons: (1) so that there are
incremental diffs to look at in the future, and also (2) the whole
original batch of changes broke SAW and I would like to isolate the
bit that causes that.