Rework the value logic, part 2: value encoding.
The plan here is to separate the internal value type (which is now
called Val) from the outward-facing raw value type, which is for now
still RawVal::T but about to be changed to Raw::T to make sure all
uses get visited.

This gives more internal and external safety, and helps prevent code
outside Values.cry from doing its own conversions and possibly getting
into trouble that way.

The original plan was to make Val private to Values.cry, but that
doesn't work: using raw_get_payload instead of a direct field
reference in to_mir_value in soroban.saw breaks saw -- or whether
it's a saw problem is debatable but it breaks the ability of saw
to match the resulting specs as overrides. So for now hold off on
that bit.