Add "some" and "none" functions to Option.cry. And a big comment.
This code hasn't made up its mind about whether it's "option" (with
constructors Some and None) or "result" (with constructors Ok and
Err).

Since it isn't capable of storing a value of a different type for the
Err case, it isn't really going to be "result", so add the proper
names for "option". Postpone cleaning it up so it isn't also using Ok
until later.