Add a proper U32 module and type to Values.cry.
The existing codebase had been using [32] as the Cryptol mapping for
both native rust u32 and also the Soroban U32Val, which was wrong (in
the sense that we want to mirror the Soroban value encoding). Use a
wrapper around [64] like the other value types.

Also add the U32Val type binding to the MIR type bundle. Do this
conditionally on a new "u32_val" feature, just in case. Should at some
point test whether it's possible to get a compiled contract without
any references to U32Val.

Also add an assume hook to go with the verify hook, which
intentionally takes the same arguments; this way for testing slow
stuff you can change "soroban.verify" to "soroban.assume" for the
pieces you aren't immediately working on without having to change
anything else around.

Using [32] instead of a real value type allowed the proofs to play
fast and loose with the distinction between u32 and U32Val in the Rust
code. On the one hand that makes the proofs simpler/easier (since e.g.
you don't have to unpack the value to do arithmetic) but it also blurs
an important distinction.

So, update the proofs to make these distinctions. Also, relatedly,
don't use the value type apparatus on plain u32 values, specificaly
to_mir_value and fresh.

The proofs do use fresh for instances of Storage::T, which is also
wrong but will need to be cleaned up separately.

Update the checked-in solver cache for sqrt.saw.