Correct the return type of soroban_value_type_to_mir for U32.
This was noted partway through the cleanup run but I didn't want to
fix it in the middle of piles of mechanical changes.

This is also probably a good time to note that tonight's commits (all
however many of them) are incrementally emptying into git cleanup that
was originally done, mostly on Friday, in a much less organized and
more exploratory manner, and that notably mostly did _not_ feature
diffs that could be read or checked. Having figured out the end goal
it was then possible to get there in a more structured and more
auditable (and in some cases more testable) way.