Add an override spec for vec_new_from_slice.
Because symbolic-length arrays don't work, this is actually six copies
of a spec specialized to lengths 0 through 5. (Ugh. But this doesn't
incur much source-level overhead.)

Note that we need this rather than vec_new_from_linear_memory (which
is the hypercall interface) because the latter takes a pointer that's
been cast to a u32, and Crucible (with some justification) refuses
that cast.

This commit depends on the fix for saw-script #2045; otherwise the
overrides fail to match correctly.