Fix the event topic consing functions.
Convenient as it might be to hide the heap state inside them, it
doesn't work; in particular they need to operate on the contract's
heap state, not the internal heap state shared by the hypercall specs.

While we could in principle declare both as ghost state, and manage
both in this module, without more structure (whether types or modules
or something else) the chances of mixing the two up are high. Also
there aren't necessarily only two in the long run and it potentially
gets very complicated.

Plus, tracking ghost state on behalf of the caller (rather than the
callee in an override spec) requires the ability to update the state,
which we don't have yet (see saw-script #2040).

So, thread the heap state through everything, ugly as it is.

There's further discussion of these points in a comment.

This commit is intentionally separate from the initial event modeling
code so that the threading-the-heap-state parts can be reverted later
if circumstances change and it becomes desirable.