Adjust how the model contents are produced in vec_new_from_slice_spec.
It's convenient to construct a fresh Cryptol var that's a sequence of
raw values of length n, and insert that into the model. But it doesn't
work: the insertion copies the values into an array, and while the
values are fine the fresh sequence itself then gets dropped, and SAW
wants to see that every fresh var you create is bound somewhere at the
end of the function.

So instead create sequences of the proper length explicitly out of the
right number of fresh vars. There doesn't seem to be a good way to do
this in a loop (and even if there were, there's a good chance it
wouldn't work anyway) so for the moment we paste the necessary number
of construction functions and explicitly call them from a big if
statement.