Assert event log equivalence rather than equality.
Provide value equality that recurses through the host heap (which is
not going to be the same as the default bit-level/SMT-level equality
for host-side objects unless the handles always end up the same.

Which they probably will for the simple examples we're doing if we're
careful about having the specs always allocate in the same order and
don't get into trying to verify the host-side code, because we're
using two copies of the same heap model that will hand out the same
handles deterministically.

But the model almost certainly doesn't hand out the same handles as
the real code, so relying on this would be wrong.

The recursive equality will short-cut if it does see the same handles,
of course.