Need to be able to assert the heap isn't full.
Things break if the heap overflows. You can go on allocating, and
there'll only actually be _real_ trouble if you allocate more than
2^32 things in a single contract run (as opposed to just arbitrarily
starting with a mostly-full heap), which nothing we're looking at
does.

However, when the heap overflows, the size wraps around. This breaks
existing handles, because the predicates checking they aren't in the
free portion of the heap fail. That causes problems.