Add a variant custom_types and an override spec for vec_unpack_to_slice.
The variant custom_types is cut down slightly so it uses a tuple struct
instead of a struct with named fields; this manifests as vectors in the
backend so we can use the existing (and tested) vector allocating code.

The additional override spec is (I think) all it's going to need to go,
but it doesn't actually work yet.