Rearrange map_unpack_to_slice_spec to fix the uniformity problem.
It now takes a list of string lengths that are the lengths of the keys
of your map. We still need a separate override spec for each
combination of lengths, but at least any given combination of lengths
is now possible.

Whole thing still doesn't work because of the Const_RefRoots thing
though.