Merge an improved version of the original too-large authorization model.
This also makes SAW run out of memory, which is not surprising since
it's strictly more complex than the other smaller versions it's based
on.

(This commit is fake in the sense that most of this logic was written
first and the other simpler versions were cut down from it. But things
are much more presentable this way.)